refactor: extract Commander registration to register-cli-commands
Move all subcommand definitions, shared credential/spinner option helpers,
and disabled hidden commands into registerCliCommands(program, deps).
index.ts keeps program setup, global options, help categories, parse/history,
and update checks. Config edit receives default credentials path via deps.

Made-with: Cursor
