+++ src/index.ts
@@ -7,43 +7,8 @@
 import 'dotenv/config';
 import process from 'node:process';
 import chalk from 'chalk';
-import { Command, Option } from 'commander';
-import { accountsCommand } from './cmds/accounts.js';
-import { balancesCommand } from './cmds/balances.js';
-import { beneficiariesCommand } from './cmds/beneficiaries.js';
-import {
-  runConfigEdit,
-  runConfigProfileDelete,
-  runConfigProfileList,
-  runConfigProfileSet,
-  runConfigProfileShow,
-} from './cmds/config-subcommands.js';
-import {
-  cardsCommand,
-  configCommand,
-  countriesCommand,
-  currenciesCommand,
-  deployCommand,
-  disableCommand,
-  enableCommand,
-  envCommand,
-  envListCommand,
-  fetchCommand,
-  logsCommand,
-  merchantsCommand,
-  newCommand,
-  publishCommand,
-  publishedCommand,
-  runCommand,
-  uploadCommand,
-  uploadEnvCommand,
-} from './cmds/index.js';
-import { payCommand } from './cmds/pay.js';
-import { simulateCommand } from './cmds/simulate.js';
-import { transactionsCommand } from './cmds/transactions.js';
-import { transferCommand } from './cmds/transfer.js';
-import { generateCompletionScript } from './completion.js';
-import { CliError, ERROR_CODES, ExitCode } from './errors.js';
+import { Command } from 'commander';
+import { registerCliCommands } from './register-cli-commands.js';
 import {
   credentialLocation,
   credentials,
@@ -54,7 +19,6 @@ import { normalizeSpinnerFlags } from './utils/spinner-flags.js';
 import {
   checkForUpdates,
   configureChalk,
-  determineExitCode,
   getVerboseMode,
   handleCliError,
   isUpdateCheckDisabled,
@@ -62,7 +26,6 @@ import {
   shouldDisplayUpdateNotification,
   showUpdateNotification,
   warnAboutSecretUsage,
-  withCommandContext,
 } from './utils.js';
 
 export { credentialLocation, credentials, optionCredentials, printTitleBox };
@@ -73,47 +36,9 @@ configureChalk();
 const version = '0.8.3';
 const program = new Command();
 
-/**
- * Commander action for commands that are registered but intentionally disabled.
- */
-function disabledCommandAction(commandName: string) {
-  return withCommandContext(commandName, async () => {
-    throw new CliError(ERROR_CODES.COMMAND_DISABLED, `The ${commandName} command is disabled.`);
-  });
-}
-
 // Improve error output for missing arguments/options
 program.showHelpAfterError();
 program.showSuggestionAfterError();
-
-// Helper for shared API credential options
-function addApiCredentialOptions(cmd: Command) {
-  return cmd
-    .option('--api-key <apiKey>', 'api key for the Investec API')
-    .option('--client-id <clientId>', 'client Id for the Investec API')
-    .option('--client-secret <clientSecret>', 'client secret for the Investec API')
-    .option('--host <host>', 'Set a custom host for the Investec Sandbox API')
-    .option('--credentials-file <credentialsFile>', 'Set a custom credentials file')
-    .option('--profile <profile>', 'Use a configuration profile (e.g., production, staging)')
-    .option(
-      '-s,--spinner',
-      'disable spinner during command execution (deprecated: use --no-spinner)'
-    )
-    .option('-v,--verbose', 'additional debugging information')
-    .option('--json', 'Output raw JSON instead of formatted table')
-    .option('--yaml', 'Output raw YAML instead of formatted table')
-    .option('--output <file>', 'Write JSON/YAML output to file instead of stdout');
-}
-
-/** Spinner + verbose only (for commands that do not use Investec API credential flags). */
-function addSpinnerVerboseOptions(cmd: Command) {
-  return cmd
-    .option(
-      '-s,--spinner',
-      'disable spinner during command execution (deprecated: use --no-spinner)'
-    )
-    .option('-v,--verbose', 'additional debugging information');
-}
 
 const spinnerFlagNormalization = normalizeSpinnerFlags(process.argv);
 process.argv = spinnerFlagNormalization.argv;
@@ -157,725 +82,7 @@ For more information about a specific command, use:
 `
   );
 
-  // Use shared options for most commands
-
-  // Card Management
-  addApiCredentialOptions(
-    program
-      .command('cards')
-      .alias('c')
-      .description(
-        'List all programmable cards. Shows card keys, numbers, and activation status for each card.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb cards
-  $ ipb cards --json
-  $ ipb cards --yaml
-  $ ipb cards --output cards.json
-  $ ipb cards --yaml --output cards.yaml
-      `
-      )
-  ).action(withCommandContext('cards', cardsCommand));
-  // Configuration
-  const configCmd = program
-    .command('config')
-    .alias('cfg')
-    .description(
-      'Configure authentication credentials. Set API keys, client credentials, and card keys for CLI operations. Supports configuration profiles for managing multiple environments.'
-    )
-    .addHelpText(
-      'after',
-      `
-Examples:
-  # Save to default credentials
-  $ ipb config --client-id <id> --client-secret <secret> --api-key <key>
-  $ ipb config --card-key <card-key>
-  
-  # Save to a profile
-  $ ipb config --profile production --client-id <id> --client-secret <secret> --api-key <key>
-  $ ipb config --profile staging --client-id <id> --client-secret <secret> --api-key <key>
-  
-  # List profiles
-  $ ipb config profile list
-  
-  # Set active profile
-  $ ipb config profile set production
-  
-  # Show active profile
-  $ ipb config profile show
-  
-  # Delete a profile
-  $ ipb config profile delete staging
-  
-  # Edit credentials in your editor
-  $ ipb config edit
-  $ ipb config edit --profile production
-      `
-    );
-
-  addApiCredentialOptions(configCmd)
-    .option('--card-key <cardKey>', 'Set your card key for the Investec API')
-    .option('--openai-key <openaiKey>', 'Set your OpenAI API key for AI code generation')
-    .option('--sandbox-key <sandboxKey>', 'Set your sandbox key for AI generation')
-    .action(withCommandContext('config', configCommand));
-
-  // Profile subcommands
-  const profileCmd = configCmd.command('profile').description('Manage configuration profiles');
-
-  profileCmd
-    .command('list')
-    .alias('ls')
-    .description('List all available configuration profiles')
-    .action(withCommandContext('config profile list', runConfigProfileList));
-
-  profileCmd
-    .command('set')
-    .description('Set the active profile (used when --profile is not specified)')
-    .argument('<profile>', 'Profile name to set as active')
-    .action(
-      withCommandContext('config profile set', (profileName: string) =>
-        runConfigProfileSet(profileName)
-      )
-    );
-
-  profileCmd
-    .command('show')
-    .description('Show the currently active profile')
-    .action(withCommandContext('config profile show', runConfigProfileShow));
-
-  profileCmd
-    .command('delete')
-    .alias('rm')
-    .description('Delete a configuration profile')
-    .argument('<profile>', 'Profile name to delete')
-    .action(
-      withCommandContext('config profile delete', (profileName: string) =>
-        runConfigProfileDelete(profileName)
-      )
-    );
-
-  // Edit command - opens credentials file in editor
-  configCmd
-    .command('edit')
-    .description(
-      'Open credentials file in your editor. Uses EDITOR environment variable or defaults to nano/vim/notepad.'
-    )
-    .option('--profile <profile>', 'Edit a specific profile instead of default credentials')
-    .addHelpText(
-      'after',
-      `
-Examples:
-  $ ipb config edit
-  $ ipb config edit --profile production
-  $ EDITOR=vim ipb config edit
-  $ EDITOR="code --wait" ipb config edit
-      `
-    )
-    .action(
-      withCommandContext('config edit', async (options: { profile?: string }) => {
-        await runConfigEdit({
-          profile: options.profile,
-          defaultCredentialsFile: credentialLocation.filename,
-        });
-      })
-    );
-
-  // Code Management
-  addApiCredentialOptions(
-    program
-      .command('deploy')
-      .alias('d')
-      .description(
-        'Deploy JavaScript code to a programmable card. Uploads code and optional environment variables, then publishes it to make it active.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb deploy -f main.js -e production -c card-123
-  $ ipb deploy -f app.js --env dev --card-key card-456
-  $ ipb deploy -f main.js -c card-789  # Deploy without environment variables
-  $ ipb deploy -f main.js -c card-123 --yes  # Skip confirmation prompt
-      `
-      )
-  )
-    .requiredOption('-f,--filename <filename>', 'JavaScript file to deploy')
-    .option('-e,--env <env>', 'Environment name (loads variables from .env.<env> file)')
-    .option('-c,--card-key <cardKey>', 'Card identifier to deploy to')
-    .option('--yes', 'Skip confirmation prompt for destructive operations')
-    .action(withCommandContext('deploy', deployCommand));
-  addApiCredentialOptions(
-    program
-      .command('logs')
-      .alias('log')
-      .description(
-        'Fetch execution logs from a card. Retrieves transaction execution logs and saves them to a JSON file.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb logs -f executions.json -c card-123
-  $ ipb logs -f logs.json --card-key card-456
-      `
-      )
-  )
-    .requiredOption('-f,--filename <filename>', 'Output filename for logs (JSON format)')
-    .option('-c,--card-key <cardKey>', 'Card identifier to fetch logs from')
-    .action(withCommandContext('logs', logsCommand));
-  addSpinnerVerboseOptions(
-    program
-      .command('run')
-      .alias('r')
-      .description(
-        'Run card code locally using the emulator. Test JavaScript code with simulated transactions without deploying to a card.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb run -f main.js -e prod --amount 60000 --currency ZAR
-  $ ipb run -f app.js --amount 10000 --merchant "Test Store" --city "Cape Town"
-      `
-      )
-  )
-    .requiredOption('-f,--filename <filename>', 'JavaScript file to execute')
-    .option('-e,--env <env>', 'Environment file to load (.env.<env>)')
-    .option('-a,--amount <amount>', 'Transaction amount in cents', '10000')
-    .option('-u,--currency <currency>', 'Currency code (ISO 4217)', 'zar')
-    .option('-z,--mcc <mcc>', 'Merchant category code', '0000')
-    .option('-m,--merchant <merchant>', 'Merchant name', 'The Coders Bakery')
-    .option('-i,--city <city>', 'Merchant city', 'Cape Town')
-    .option('-o,--country <country>', 'Country code (ISO 3166-1 alpha-2)', 'ZA')
-    .action(withCommandContext('run', runCommand));
-  addApiCredentialOptions(
-    program
-      .command('fetch')
-      .alias('f')
-      .description(
-        'Fetch saved code from a card. Downloads the code currently saved on a card and saves it to a local file.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb fetch -f saved-code.js -c card-123
-  $ ipb fetch -f backup.js --card-key card-456
-      `
-      )
-  )
-    .requiredOption('-f,--filename <filename>', 'Local filename to save the code to')
-    .option('-c,--card-key <cardKey>', 'Card identifier to fetch code from')
-    .action(withCommandContext('fetch', fetchCommand));
-  addApiCredentialOptions(
-    program
-      .command('upload')
-      .alias('up')
-      .description(
-        'Upload code to a card without publishing. Saves JavaScript code to a card but does not activate it. Use publish command to activate.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb upload -f main.js -c card-123
-  $ ipb upload -f app.js --card-key card-456
-      `
-      )
-  )
-    .requiredOption('-f,--filename <filename>', 'JavaScript file to upload')
-    .option('-c,--card-key <cardKey>', 'Card identifier to upload code to')
-    .action(withCommandContext('upload', uploadCommand));
-  // Environment Management
-  addApiCredentialOptions(
-    program
-      .command('env')
-      .description(
-        'Download environment variables from a card. Retrieves all environment variables configured on a card and saves them to a JSON file.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb env -f env.json -c card-123
-  $ ipb env -f variables.json --card-key card-456
-      `
-      )
-  )
-    .requiredOption(
-      '-f,--filename <filename>',
-      'Output filename for environment variables (JSON format)'
-    )
-    .option('-c,--card-key <cardKey>', 'Card identifier to fetch environment from')
-    .action(withCommandContext('env', envCommand));
-  addApiCredentialOptions(
-    program
-      .command('env-list')
-      .description(
-        'List all supported environment variables. Shows available environment variables with descriptions, usage examples, and default values.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb env-list
-  $ ipb env-list --json
-  $ ipb env-list --yaml --output env-vars.yaml
-      `
-      )
-  ).action(withCommandContext('env-list', envListCommand));
-  addApiCredentialOptions(
-    program
-      .command('upload-env')
-      .description(
-        'Upload environment variables to a card. Reads environment variables from a JSON file and uploads them to a card.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb upload-env -f env.json -c card-123
-  $ ipb upload-env -f variables.json --card-key card-456
-      `
-      )
-  )
-    .requiredOption('-f,--filename <filename>', 'JSON file containing environment variables')
-    .option('-c,--card-key <cardKey>', 'Card identifier to upload environment to')
-    .action(withCommandContext('upload-env', uploadEnvCommand));
-  addApiCredentialOptions(
-    program
-      .command('published')
-      .description(
-        'Download the currently published code from a card. Retrieves the active code currently running on a card and saves it to a local file.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb published -f published-code.js -c card-123
-  $ ipb published -f active.js --card-key card-456
-      `
-      )
-  )
-    .requiredOption('-f,--filename <filename>', 'Local filename to save the published code to')
-    .option('-c,--card-key <cardKey>', 'Card identifier to fetch published code from')
-    .action(withCommandContext('published', publishedCommand));
-  addApiCredentialOptions(
-    program
-      .command('publish')
-      .alias('pub')
-      .description(
-        'Publish previously uploaded code to make it active. Activates code that was uploaded using the upload command. Requires code ID from upload.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb publish -f main.js --code-id code-123 -c card-456
-  $ ipb publish -f app.js -i code-789 --card-key card-123
-  $ ipb publish -f main.js --code-id code-123 -c card-456 --yes  # Skip confirmation
-      `
-      )
-  )
-    .requiredOption(
-      '-f,--filename <filename>',
-      'JavaScript file to publish (must match uploaded code)'
-    )
-    .requiredOption('-i,--code-id <codeId>', 'Code ID from previous upload command')
-    .option('-c,--card-key <cardKey>', 'Card identifier to publish code to')
-    .option('--yes', 'Skip confirmation prompt for destructive operations')
-    .action(withCommandContext('publish', publishCommand));
-  addApiCredentialOptions(
-    program
-      .command('simulate')
-      .description(
-        'Test code using the online simulator. Runs JavaScript code in the Investec cloud environment with simulated transactions. Similar to run but uses cloud infrastructure.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb simulate -f main.js -c card-123 --amount 60000 --currency ZAR
-  $ ipb simulate -f app.js --card-key card-456 --env production
-      `
-      )
-  )
-    .requiredOption('-f,--filename <filename>', 'JavaScript file to simulate (required)')
-    .option('-c,--card-key <cardKey>', 'Card identifier for simulation')
-    .option('-e,--env <env>', 'Environment name', 'development')
-    .option('-a,--amount <amount>', 'Transaction amount in cents', '10000')
-    .option('-u,--currency <currency>', 'Currency code (ISO 4217)', 'zar')
-    .option('-z,--mcc <mcc>', 'Merchant category code', '0000')
-    .option('-m,--merchant <merchant>', 'Merchant name', 'The Coders Bakery')
-    .option('-i,--city <city>', 'Merchant city', 'Cape Town')
-    .option('-o,--country <country>', 'Country code (ISO 3166-1 alpha-2)', 'ZA')
-    .action(withCommandContext('simulate', simulateCommand));
-  addApiCredentialOptions(
-    program
-      .command('enable')
-      .description(
-        'Enable programmable code on a card. Activates programmable card functionality. Code must be deployed and published first.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb enable -c card-123
-  $ ipb enable --card-key card-456
-      `
-      )
-  )
-    .option('-c,--card-key <cardKey>', 'Card identifier to enable code on')
-    .action(withCommandContext('enable', enableCommand));
-  addApiCredentialOptions(
-    program
-      .command('disable')
-      .description(
-        'Disable programmable code on a card. Deactivates programmable card functionality. Code remains deployed but will not execute on transactions.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb disable -c card-123
-  $ ipb disable --card-key card-456
-  $ ipb disable -c card-123 --yes  # Skip confirmation
-      `
-      )
-  )
-    .option('-c,--card-key <cardKey>', 'Card identifier to disable code on')
-    .option('--yes', 'Skip confirmation prompt for destructive operations')
-    .action(withCommandContext('disable', disableCommand));
-  // Reference Data
-  addApiCredentialOptions(
-    program
-      .command('currencies')
-      .description(
-        'List all supported currency codes. Shows ISO 4217 currency codes and names available for use in transaction simulations and operations.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb currencies
-  $ ipb currencies --json
-  $ ipb currencies --yaml
-  $ ipb currencies --output currencies.json
-      `
-      )
-  ).action(withCommandContext('currencies', currenciesCommand));
-  addApiCredentialOptions(
-    program
-      .command('countries')
-      .description(
-        'List all supported country codes. Shows ISO 3166-1 alpha-2 country codes and names available for use in transaction simulations.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb countries
-  $ ipb countries --json
-  $ ipb countries --yaml
-  $ ipb countries --output countries.json
-      `
-      )
-  ).action(withCommandContext('countries', countriesCommand));
-  addApiCredentialOptions(
-    program
-      .command('merchants')
-      .description(
-        'List merchant categories and codes. Shows merchant category codes (MCC) with descriptions for use in transaction simulations.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb merchants
-  $ ipb merchants --json
-  $ ipb merchants --yaml
-  $ ipb merchants --output merchants.json
-      `
-      )
-  ).action(withCommandContext('merchants', merchantsCommand));
-  // Account Management
-  addApiCredentialOptions(
-    program
-      .command('accounts')
-      .alias('acc')
-      .description(
-        'List all Investec accounts. Shows account IDs, account numbers, product names, and reference names for all linked accounts.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb accounts
-  $ ipb accounts --json
-  $ ipb accounts --yaml
-  $ ipb accounts --output accounts.json
-      `
-      )
-  ).action(withCommandContext('accounts', accountsCommand));
-  addApiCredentialOptions(
-    program
-      .command('balances')
-      .alias('bal')
-      .description(
-        'Get account balance information. Retrieves current balance, available balance, and budget balance for a specific account.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb balances <accountId>
-  $ ipb balances acc-123 --json
-  $ ipb balances acc-123 --yaml
-  $ ipb balances acc-123 --output balance.json
-      `
-      )
-  )
-    .argument('accountId', 'Account ID to fetch balances for')
-    .action(withCommandContext('balances', balancesCommand));
-  // Payments
-  addApiCredentialOptions(
-    program
-      .command('transfer')
-      .description(
-        'Transfer money between your own accounts. Moves funds from one account to another. Prompts interactively for missing information.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb transfer <accountId> <beneficiaryAccountId> <amount> <reference>
-  $ ipb transfer acc-123 acc-456 100.50 "Payment for services"
-  $ ipb transfer acc-123 acc-456 100.50 "Payment" --yes  # Skip confirmation
-      `
-      )
-  )
-    .argument('accountId', 'Account ID to transfer from')
-    .argument('beneficiaryAccountId', 'Beneficiary account ID to transfer to')
-    .argument('amount', 'Amount to transfer in rands (e.g. 100.00)')
-    .argument('reference', 'Payment reference message')
-    .option('--yes', 'Skip confirmation prompt for destructive operations')
-    .action(withCommandContext('transfer', transferCommand));
-  addApiCredentialOptions(
-    program
-      .command('pay')
-      .description(
-        'Pay a beneficiary from your account. Makes a payment to a registered beneficiary. Requires confirmation before executing.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb pay <accountId> <beneficiaryId> <amount> <reference>
-  $ ipb pay acc-123 ben-456 250.00 "Monthly subscription"
-  $ ipb pay acc-123 ben-456 250.00 "Payment" --yes  # Skip confirmation
-      `
-      )
-  )
-    .argument('accountId', 'Account ID to pay from')
-    .argument('beneficiaryId', 'Beneficiary ID to pay to')
-    .argument('amount', 'Amount to pay in rands (e.g. 100.00)')
-    .argument('reference', 'Payment reference message')
-    .option('--yes', 'Skip confirmation prompt for destructive operations')
-    .action(withCommandContext('pay', payCommand));
-  addApiCredentialOptions(
-    program
-      .command('transactions')
-      .alias('tx')
-      .description(
-        'Get transaction history for an account. Retrieves and displays recent transactions with full details including amounts, dates, and merchants.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb transactions <accountId>
-  $ ipb transactions acc-123 --json
-  $ ipb transactions acc-123 --yaml
-  $ ipb transactions acc-123 --output transactions.json
-      `
-      )
-  )
-    .argument('accountId', 'Account ID to fetch transactions for')
-    .action(withCommandContext('transactions', transactionsCommand));
-  addApiCredentialOptions(
-    program
-      .command('beneficiaries')
-      .description(
-        'List all beneficiaries. Shows all registered beneficiaries linked to your Investec profile with their IDs and details.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb beneficiaries
-  $ ipb beneficiaries --json
-  $ ipb beneficiaries --yaml
-  $ ipb beneficiaries --output beneficiaries.json
-      `
-      )
-  ).action(withCommandContext('beneficiaries', beneficiariesCommand));
-  // AI & Code Generation
-  addSpinnerVerboseOptions(
-    program
-      .command('new')
-      .description(
-        'Create a new project with scaffolding. Generates a new project directory with template files and directory structure for card code development.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb new my-project
-  $ ipb new my-app --template petro
-  $ ipb new my-project --force
-      `
-      )
-  )
-    .argument('name', 'Project name (will create a directory with this name)')
-    .option('--force', 'Overwrite existing project directory if it exists')
-    .addOption(
-      new Option('--template <template>', 'Template to use for project structure')
-        .default('default')
-        .choices(['default', 'petro'])
-    )
-    .action(withCommandContext('new', newCommand));
-  addSpinnerVerboseOptions(
-    program
-      .command('ai', { hidden: true })
-      .description(
-        'Generate programmable card code using AI. Creates JavaScript code for programmable cards from natural language descriptions using OpenAI or sandbox service.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb ai "block transactions over R1000"
-  $ ipb ai "only allow transactions at Woolworths" -f rules.js
-  $ ipb ai "send SMS notification after payment" --force
-      `
-      )
-  )
-    .argument('prompt', 'Natural language description of the card code behavior')
-    .option('-f,--filename <filename>', 'Output filename for generated code', 'ai-generated.js')
-    .option('--force', 'Overwrite existing file if it exists')
-    .action(disabledCommandAction('ai'));
-  addSpinnerVerboseOptions(
-    program
-      .command('bank', { hidden: true })
-      .description(
-        'Use AI to interact with your bank account. Performs banking operations using natural language prompts. Uses AI to interpret requests and execute appropriate banking functions.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb bank "Show me my last 5 transactions"
-  $ ipb bank "What is my account balance?"
-  $ ipb bank "Transfer R100 to account acc-123"
-      `
-      )
-      .argument('prompt', 'Natural language description of the banking operation to perform')
-  ).action(disabledCommandAction('bank'));
-  // Authentication (disabled — hidden commands for backward compatibility with scripts)
-  addSpinnerVerboseOptions(
-    program
-      .command('register', { hidden: true })
-      .description(
-        'Register for the sandbox AI service. Creates an account for using AI code generation features without requiring your own OpenAI API key.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb register -e user@example.com -p securepassword
-  $ ipb register --email user@example.com --password securepassword
-      
-Note: After registration, message in #12_sandbox-playground with your email to activate your account.
-      `
-      )
-      .option('-e,--email <email>', 'Email address for registration')
-      .option('-p,--password <password>', 'Password for your account')
-  ).action(disabledCommandAction('register'));
-  addSpinnerVerboseOptions(
-    program
-      .command('login', { hidden: true })
-      .description(
-        'Login to the sandbox AI service. Authenticates with the sandbox service and saves access token for use with AI generation commands.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb login -e user@example.com -p securepassword
-  $ ipb login --email user@example.com --password securepassword
-      `
-      )
-      .option('-e,--email <email>', 'Your registered email address')
-      .option('-p,--password <password>', 'Your account password')
-  ).action(disabledCommandAction('login'));
-  // Utilities
-  program
-    .command('completion')
-    .description(
-      'Generate shell completion script. Creates autocomplete scripts for bash or zsh to enable tab completion for commands and options.'
-    )
-    .addHelpText(
-      'after',
-      `
-Examples:
-  $ ipb completion bash > /etc/bash_completion.d/ipb
-  $ ipb completion zsh > ~/.zsh/completions/_ipb
-  $ source <(ipb completion bash)  # For current session only
-      `
-    )
-    .argument('<shell>', 'Shell type (bash or zsh)')
-    .action((shell) => {
-      try {
-        const completionScript = generateCompletionScript(shell);
-        console.log(completionScript);
-      } catch (error) {
-        const errorMessage = error instanceof Error ? error.message : String(error);
-        console.error(chalk.red(`Error generating completion script: ${errorMessage}`));
-        const exitCode =
-          error instanceof CliError
-            ? determineExitCode(error, error.code, errorMessage)
-            : ExitCode.GENERAL_ERROR;
-        process.exit(exitCode);
-      }
-    });
-
-  // Documentation generation command
-  addSpinnerVerboseOptions(
-    program
-      .command('docs')
-      .description(
-        'Generate command documentation. Creates markdown documentation from CLI command definitions and writes it to GENERATED_README.md.'
-      )
-      .addHelpText(
-        'after',
-        `
-Examples:
-  $ ipb docs
-  $ ipb docs --output COMMANDS.md
-      `
-      )
-  )
-    .option('--output <file>', 'Output file path (default: GENERATED_README.md)')
-    .action(async (options: { output?: string; verbose?: boolean; spinner?: boolean }) => {
-      const { docsCommand } = await import('./cmds/docs.js');
-      await docsCommand(options.output || 'GENERATED_README.md', program, {
-        verbose: options.verbose,
-        spinner: options.spinner,
-      });
-    });
+  registerCliCommands(program, { credentialLocation });
 
   // Check for --check-updates flag in raw arguments
   const hasCheckUpdatesFlag = process.argv.includes('--check-updates');
